Create draft of Readme.md
Included a couple references to the original paper, the original article and an excerpt of lecture notes from MIT's 6.824 (Paxos made simple). I hope to add more details as to the working of Paxos, including a layman description of the problem and a concise and efficient solution that can be easily deployed.